Resolve extension merge conflicts
